requirements.txt
gunicorn==0.11.2
django-compressor==1.3
django-summernote==0.5.8
django-bootstrap3
